refactor: 테스트 코드 리팩토링
- mock 디렉토리 위치 변경
- 테스트를 위한 queryClient 별도 생성
- 프로덕션 코드 변경에 따른 테스트 코드 변경
