+++ __test__/mocks/queryClient.ts
@@ -1,0 +1,8 @@
+import { QueryClient } from '@tanstack/react-query';
+
+export const queryClient = new QueryClient({
+  defaultOptions: {
+    queries: { retry: 0 },
+    mutations: { retry: 0 },
+  },
+});
