+++ vitest.setup.ts
@@ -1,5 +1,5 @@
 import { beforeAll, afterEach, afterAll } from 'vitest';
-import server from './src/mocks/server';
+import server from './__test__/mocks/server';
 
 beforeAll(() => {
   server.listen({ onUnhandledRequest: 'error' });
